feat: Extract panic recovery interceptor to shared library
Move the gRPC panic recovery interceptors from position-keeping to
shared/pkg/interceptors so all services can use consistent panic
handling. This prevents server crashes from unhandled panics and
ensures clients receive proper gRPC error codes.

Changes:
- Create shared/pkg/interceptors with RecoveryUnaryInterceptor,
  RecoveryStreamInterceptor, and RecoveryStreamInterceptorWithWrappedStream
- Add recovery interceptors to current-account, financial-accounting,
  and payment-order services
- Update position-keeping to import from shared instead of local package
- Remove duplicate interceptors package from position-keeping

Closes #224
